Server_Modify_Room Class
Room Init함수 버그 수정!

+++ NeoSports_Server/Server_BOOST/ServerMain/GameMG.h
@@ -7,7 +7,7 @@ class GameMG
 private:
 	GAME_INDEX _curGame;
 
-	int _ropePos;
+	float _ropePos;
 
 public:
 	GameMG();
@@ -17,6 +17,6 @@ public:
 
 	void Init();
 
-	void SetRopePos(int ropePos);
+	void SetRopePos(float ropePos);
 	int GetRopePos();
 };
